src/modules/client-adm/facade/client-adm.facade.spec.ts
import { Sequelize } from "sequelize-typescript";
import { ClientModel } from "../repository/client.model";
import { ClientRepository } from "../repository/client.repository";
import { AddClientUseCase } from "../usecase/add-client/add-client.usecase";
import { ClientAdmFacade } from "./client-adm.facade";

describe("Client Adm Facade test", () => {
  let sequelize: Sequelize;

  beforeEach(async () => {
    sequelize = new Sequelize({
      dialect: "sqlite",
      storage: ":memory:",
      logging: false,
      sync: { force: true },
    });

    await sequelize.addModels([ClientModel]);
    await sequelize.sync();
  });

  afterEach(async () => {
    await sequelize.close();
  });

  it("should create a client", async () => {
    const repository = new ClientRepository();
    const addClientUseCase = new AddClientUseCase(repository);
    const clientFacade = new ClientAdmFacade({
      addClientUseCase: addClientUseCase,
      findClientUseCase: undefined,
    });

    const input = {
      id: "1",
      name: "Milena",
      email: "milena@mail.com",
      address: "Address 1",
    };

    await clientFacade.add(input);

    const client = await ClientModel.findOne({ where: { id: "1" } });

    expect(client).toBeDefined();
    expect(client.name).toBe(input.name);
    expect(client.email).toBe(input.email);
    expect(client.address).toBe(input.address);
  });
});
